+++ azure/CHANGELOG.md
@@ -4,7 +4,11 @@ This file documents changes to the `workbench-azure` library, including notes on
 
 ## 0.4
 
-SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-azure" % "0.4-c41881f"`
+SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-azure" % "0.4-TRAVIS-REPLACE-ME"`
+
+### Added
+
+- Added `AzureRelayService#getHybridConnectionKey`
 
 ## 0.3